Feat: 레포 분석 저장 기능 완성
saveAnalysis api 완성

+++ backend/src/main/java/com/dev101/coa/domain/repo/entity/RepoView.java
@@ -1,5 +1,6 @@
 package com.dev101.coa.domain.repo.entity;
 
+import com.dev101.coa.domain.code.entity.Code;
 import com.dev101.coa.domain.member.entity.Member;
 import com.dev101.coa.global.common.BaseEntity;
 import jakarta.persistence.*;
@@ -35,7 +36,7 @@ public class RepoView extends BaseEntity {
     @Column(columnDefinition = "TEXT")
     private String repoViewResult;
 
-    private Integer repoViewCommitCnt;
+    private Long repoViewCommitCnt;
 
     private Integer repoViewPrCnt;
